Disable use of cached .reb files if they exist
Reia drops .reb files onto the filesystem whenever it loads a .re file
(similar to how Rubinius drops .rbc files which contain compiled Ruby
code). This is a great performance improvement becasue it means we don't
need to run the Reia compiler unless a particular source code file has changed.

However, if we're making major changes to things like the parser (like, say,
switching it to a PEG) this means the parser won't get exercised because
Reia is using the existing .reb files instead.

I've added a USE_BIN_CACHE macro to the top of reia_internal.erl which allows
an easy way of toggling this behavior on and off depending on if it's desired.